Remove the Local TSO test cases
