--- .github/workflows/tso-consistency-test.yaml
@@ -1,17 +1,0 @@
-name: TSO Consistency Test
-on:
-  # Only run when the new code is merged into master.
-  push:
-    branches:
-      - master
-jobs:
-  tso-consistency-test:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-      - uses: actions/setup-go@v5
-        with:
-          go-version: '1.23'
-      - name: Make TSO Consistency Test
-        run: make test-tso-consistency
